Removing unsafe fprintf() call from signal handler.
There are only a very limited selection of safe functions
that should be called from signal handlers, and unfortunately
fprintf isn't one of them (the "streams" may have internal
locks, etc.).

I'm leaving the fubar message however, since it should never
happen, and if it does I'd rather the chance of getting some
information out than not (it's not in a pretty state anyways).